回归修补进度快照（第 8 轮前）：md 3324→1705，stun/channel/bridge 归零
- 13 服务 7 轮并行回归：总计 md 从 3324 降至 1705（权威口径：先 manifest 重新分类再全量重生成）
- stun/channel/bridge 三个服务非 identical 归零
- 已验证源码形态技巧：continue/goto/switch 单 case/register/无符号常量/
  bool 局部/__builtin_alloca/受控内联 asm（复现编译器折叠的死比较与 nop 落地）
- 工具链：c6446r/c6/c5 变体定稿；SRC_OPTS 多选项 bug 修复；report_resolve
  归一化增强与跳转表误判修复；pthread_equal 豁免
- 全部脚本 /mnt/d 硬编码迁移到当前工作区
- dbmw 第 6 轮 DBManager.cpp 清空事故已恢复（恢复 HEAD + 补 ~STGuildCargoLog）
- 进度文档：docs/regression_progress_2026-08-11.md 全量汇总

+++ dnf_decompile/source/ChannelOld/DNFChannelServer/TCPUser.cpp
@@ -73,7 +73,7 @@ void ChannelServiceApp::TCPUser::onError()
 
 void ChannelServiceApp::TCPUser::onClose(char* file, int line)
 {
-    GLOG(gFileLogInfo, "call onClose from " << file << ", " << line << ", TCPUSER=" << getACCID() << ", error=" << strerror(*__errno_location()) << ", dis=" << bDisconnected_);
+    GLOG(gFileLogInfo, "call onClose from " << file << ", " << line << ", TCPUSER=" << (int)getACCID() << ", error=" << strerror(*__errno_location()) << ", dis=" << bDisconnected_);
     if (bDisconnected_)
         return;
     TManager<ChannelService>::getManager()->UserPools::destroyTCPUser(this, "TCPUser.cpp", 0x6f);
@@ -86,7 +86,7 @@ void ChannelServiceApp::TCPUser::onClose(char* file, int line)
 
 void ChannelServiceApp::TCPUser::onRead(char* file, int line)
 {
-    GLOG(gFileLogInfo, "OnRead Call from  " << file << ", line " << line << ", TCPUSER=" << getACCID());
+    GLOG(gFileLogInfo, "OnRead Call from  " << file << ", line " << line << ", TCPUSER=" << (int)getACCID());
     if (pSock_->getHandle() < 0)
         return;
     if (isAboutToDisconnect() || isDisconnected())
